Add support for Simple Features
Supports `sf`, `sfc` and `sfg` class objects. `GEOMETRY` and
`GEOMETRYCOLLECTION` are not supported since they are not compatible
with shapefiles, and introduces ambiguities for QGIS modules.

If `load_output` is not `NULL`, and QGIS output is a vector layer,
`run_qgis` returns an object of the same class as the vector input.

`Spatial*` and `sf` objects can be supplied in the same parameter list.
In that case, a warning is given, and an `sf` object is returned (given
that `load_output` is not `NULL`)